move the code from bread/layout/error.py to bread/views/error.py and change the href to homepage to be the more appropriate one.
